+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.67.2] - 2026-05-04
+
 ### Changed
 - Forward `VtexRCSessionIdv7` and `VtexRCMacIdv7` cookies to the Checkout API.